Move `OptionalRangeInclusive` to `query`
